Moved app important files to /app folder
